+++ Blaze3SDK/Blaze/GameManager/ReplicatedGameData.cs
@@ -21,7 +21,7 @@ namespace Blaze3SDK.Blaze.GameManager
 		public SortedDictionary<string, string> mEntryCriteriaMap;
 
 		[TdfMember("GID")]
-		public uint mGameId;
+		public ulong mGameId;
 
 		[TdfMember("GNAM")]
 		public string mGameName;
